requirements.txt
numpy==1.19.5
pandas==1.2.4
tensorflow==2.4.1
torch==1.8.1
scikit-learn==0.24.2
librosa==0.8.0
opencv-python==4.5.1.48
matplotlib==3.3.4
seaborn==0.11.1
nltk==3.5
spacy==3.0.6
joblib==1.0.1
flask==1.1.2
gunicorn==20.1.0
psycopg2-binary==2.8.6
python-dotenv==0.17.0


